Do not run CI for PRs where both branches are from the same repo
When a pull request is created by comparing branches that are on the same repo, duplicate CI runs would be created: one for the branch push, and one for the pull request.
This commit prevents that, whilst preserving the ability to run the CI for each branch push to a fork, and for each pull request originating from a fork.