fixed generated call to ValueType.GetHashCode
The bug (https://github.com/m4rs-mt/ILGPU/issues/614) occured when a type inside SpecializedValue does not implement GetHashCode, and instead inherits one from ValueType.

Because ValueType itself is not a struct, to call its GetHashCode the first parameter must be an object reference. Therefore the object of type T (which SpecializedValue requires to be a struct) must be boxed before calling GetHashCode.