Interop debugging .IOOperations() clarification
Added a clarification to the PrintF-Like Debugging. Interop.Write() and Interop.WriteLine() do nothing if the Accelerator is not built with the .IOOperations() builder function. This is not well clarified in the sample nor the docs, so I think it's a worthwhile contribution.

I copied a bunch of the wording from some of the other notes above. But since this directly pertains to the last section I thought it would make most sense down there. Let me know if I should make any changes to what I have added or if it could be more concise or accurate.